feat: add quick copy function
